More files moved to common folder
