--- src/Microsoft.Data.SqlClient/netfx/src/Microsoft/Data/SqlClient/OnChangedEventHandler.cs
@@ -1,10 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace Microsoft.Data.SqlClient
-{
-    /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/OnChangeEventHandler.xml' path='docs/members[@name="OnChangeEventHandler"]/*'/>
-    public delegate void OnChangeEventHandler(object sender, SqlNotificationEventArgs e);
-}
-
