--- src/Microsoft.Data.SqlClient/netfx/src/Microsoft/Data/SqlClient/SqlConnectionPoolProviderInfo.cs
@@ -1,25 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using Microsoft.Data.ProviderBase;
-
-namespace Microsoft.Data.SqlClient
-{
-    internal sealed class SqlConnectionPoolProviderInfo : DbConnectionPoolProviderInfo
-    {
-        private string _instanceName;
-
-        internal string InstanceName
-        {
-            get
-            {
-                return _instanceName;
-            }
-            set
-            {
-                _instanceName = value;
-            }
-        }
-    }
-}
