--- src/Microsoft.Data.SqlClient/netfx/src/Microsoft/Data/SqlClient/SqlRowUpdatingEventHandler.cs
@@ -1,9 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace Microsoft.Data.SqlClient
-{
-    /// <include file='..\..\..\..\..\..\..\doc\snippets\Microsoft.Data.SqlClient\SqlRowUpdatingEventHandler.xml' path='docs/members[@name="SqlRowUpdatingEventHandler"]/SqlRowUpdatingEventHandler/*' />
-    public delegate void SqlRowUpdatingEventHandler(object sender, SqlRowUpdatingEventArgs e);
-}
