+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/OnChangedEventHandler.cs
@@ -4,6 +4,7 @@
 
 namespace Microsoft.Data.SqlClient
 {
-    /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/OnChangeEventHandler.xml' path='docs/members[@name="OnChangeEventHandler"]/*'/>
+    /// <include file='../../../../../../doc/snippets/Microsoft.Data.SqlClient/OnChangeEventHandler.xml' path='docs/members[@name="OnChangeEventHandler"]/*'/>
     public delegate void OnChangeEventHandler(object sender, SqlNotificationEventArgs e);
 }
+
