+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/SqlInfoMessageEventHandler.cs
@@ -4,6 +4,6 @@
 
 namespace Microsoft.Data.SqlClient
 {
-    /// <include file='..\..\..\..\..\..\..\doc\snippets\Microsoft.Data.SqlClient\SqlInfoMessageEventHandler.xml' path='docs/members[@name="SqlInfoMessageEventHandler"]/SqlInfoMessageEventHandler/*' />
+    /// <include file='..\..\..\..\..\..\doc\snippets\Microsoft.Data.SqlClient\SqlInfoMessageEventHandler.xml' path='docs/members[@name="SqlInfoMessageEventHandler"]/SqlInfoMessageEventHandler/*' />
     public delegate void SqlInfoMessageEventHandler(object sender, SqlInfoMessageEventArgs e);
 }
