+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/SqlRowUpdatingEventHandler.cs
@@ -4,6 +4,6 @@
 
 namespace Microsoft.Data.SqlClient
 {
-    /// <include file='..\..\..\..\..\..\..\doc\snippets\Microsoft.Data.SqlClient\SqlRowUpdatingEventHandler.xml' path='docs/members[@name="SqlRowUpdatingEventHandler"]/SqlRowUpdatingEventHandler/*' />
+    /// <include file='..\..\..\..\..\..\doc\snippets\Microsoft.Data.SqlClient\SqlRowUpdatingEventHandler.xml' path='docs/members[@name="SqlRowUpdatingEventHandler"]/SqlRowUpdatingEventHandler/*' />
     public delegate void SqlRowUpdatingEventHandler(object sender, SqlRowUpdatingEventArgs e);
 }
